"button" renamed to "gui"
