+++ src/gui.cpp
@@ -1,4 +1,4 @@
-#include "button.hpp"
+#include "gui.hpp"
 
 InputManager::InputManager()
 {
